+++ Demo/Demo/UI/Configuration.swift
@@ -1,0 +1,181 @@
+//
+//  Configuration.swift
+//  Demo
+//
+//  Created by Davide De Rosa on 6/13/20.
+//  Copyright (c) 2023 Davide De Rosa. All rights reserved.
+//
+//  https://github.com/keeshux
+//
+//  This file is part of TunnelKit.
+//
+//  TunnelKit is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  TunnelKit is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU General Public License for more details.
+//
+//  You should have received a copy of the GNU General Public License
+//  along with TunnelKit.  If not, see <http://www.gnu.org/licenses/>.
+//
+
+import Foundation
+import TunnelKitCore
+import TunnelKitOpenVPN
+import TunnelKitWireGuard
+
+#if os(macOS)
+let appGroup = "DTDYD63ZX9.group.com.algoritmico.TunnelKit.Demo"
+private let bundleComponent = "macos"
+#elseif os(iOS)
+let appGroup = "group.com.algoritmico.TunnelKit.Demo"
+private let bundleComponent = "ios"
+#else
+let appGroup = "group.com.algoritmico.TunnelKit.Demo"
+private let bundleComponent = "tvos"
+#endif
+
+enum TunnelIdentifier {
+    static let openVPN = "com.algoritmico.\(bundleComponent).TunnelKit.Demo.OpenVPN-Tunnel"
+
+    static let wireGuard = "com.algoritmico.\(bundleComponent).TunnelKit.Demo.WireGuard-Tunnel"
+}
+
+extension OpenVPN {
+    struct DemoConfiguration {
+        static let ca = OpenVPN.CryptoContainer(pem: """
+-----BEGIN CERTIFICATE-----
+MIIFozCCA4ugAwIBAgIBATANBgkqhkiG9w0BAQ0FADBAMQswCQYDVQQGEwJDSDEV
+MBMGA1UEChMMUHJvdG9uVlBOIEFHMRowGAYDVQQDExFQcm90b25WUE4gUm9vdCBD
+QTAeFw0xNzAyMTUxNDM4MDBaFw0yNzAyMTUxNDM4MDBaMEAxCzAJBgNVBAYTAkNI
+MRUwEwYDVQQKEwxQcm90b25WUE4gQUcxGjAYBgNVBAMTEVByb3RvblZQTiBSb290
+IENBMIICIjANBgkqhkiG9w0BAQEFAAOCAg8AMIICCgKCAgEAt+BsSsZg7+AuqTq7
+vDbPzfygtl9f8fLJqO4amsyOXlI7pquL5IsEZhpWyJIIvYybqS4s1/T7BbvHPLVE
+wlrq8A5DBIXcfuXrBbKoYkmpICGc2u1KYVGOZ9A+PH9z4Tr6OXFfXRnsbZToie8t
+2Xjv/dZDdUDAqeW89I/mXg3k5x08m2nfGCQDm4gCanN1r5MT7ge56z0MkY3FFGCO
+qRwspIEUzu1ZqGSTkG1eQiOYIrdOF5cc7n2APyvBIcfvp/W3cpTOEmEBJ7/14RnX
+nHo0fcx61Inx/6ZxzKkW8BMdGGQF3tF6u2M0FjVN0lLH9S0ul1TgoOS56yEJ34hr
+JSRTqHuar3t/xdCbKFZjyXFZFNsXVvgJu34CNLrHHTGJj9jiUfFnxWQYMo9UNUd4
+a3PPG1HnbG7LAjlvj5JlJ5aqO5gshdnqb9uIQeR2CdzcCJgklwRGCyDT1pm7eoiv
+WV19YBd81vKulLzgPavu3kRRe83yl29It2hwQ9FMs5w6ZV/X6ciTKo3etkX9nBD9
+ZzJPsGQsBUy7CzO1jK4W01+u3ItmQS+1s4xtcFxdFY8o/q1zoqBlxpe5MQIWN6Qa
+lryiET74gMHE/S5WrPlsq/gehxsdgc6GDUXG4dk8vn6OUMa6wb5wRO3VXGEc67IY
+m4mDFTYiPvLaFOxtndlUWuCruKcCAwEAAaOBpzCBpDAMBgNVHRMEBTADAQH/MB0G
+A1UdDgQWBBSDkIaYhLVZTwyLNTetNB2qV0gkVDBoBgNVHSMEYTBfgBSDkIaYhLVZ
+TwyLNTetNB2qV0gkVKFEpEIwQDELMAkGA1UEBhMCQ0gxFTATBgNVBAoTDFByb3Rv
+blZQTiBBRzEaMBgGA1UEAxMRUHJvdG9uVlBOIFJvb3QgQ0GCAQEwCwYDVR0PBAQD
+AgEGMA0GCSqGSIb3DQEBDQUAA4ICAQCYr7LpvnfZXBCxVIVc2ea1fjxQ6vkTj0zM
+htFs3qfeXpMRf+g1NAh4vv1UIwLsczilMt87SjpJ25pZPyS3O+/VlI9ceZMvtGXd
+MGfXhTDp//zRoL1cbzSHee9tQlmEm1tKFxB0wfWd/inGRjZxpJCTQh8oc7CTziHZ
+ufS+Jkfpc4Rasr31fl7mHhJahF1j/ka/OOWmFbiHBNjzmNWPQInJm+0ygFqij5qs
+51OEvubR8yh5Mdq4TNuWhFuTxpqoJ87VKaSOx/Aefca44Etwcj4gHb7LThidw/ky
+zysZiWjyrbfX/31RX7QanKiMk2RDtgZaWi/lMfsl5O+6E2lJ1vo4xv9pW8225B5X
+eAeXHCfjV/vrrCFqeCprNF6a3Tn/LX6VNy3jbeC+167QagBOaoDA01XPOx7Odhsb
+Gd7cJ5VkgyycZgLnT9zrChgwjx59JQosFEG1DsaAgHfpEl/N3YPJh68N7fwN41Cj
+zsk39v6iZdfuet/sP7oiP5/gLmA/CIPNhdIYxaojbLjFPkftVjVPn49RqwqzJJPR
+N8BOyb94yhQ7KO4F3IcLT/y/dsWitY0ZH4lCnAVV/v2YjWAWS3OWyC8BFx/Jmc3W
+DK/yPwECUcPgHIeXiRjHnJt0Zcm23O2Q3RphpU+1SO3XixsXpOVOYP6rJIXW9bMZ
+A1gTTlpi7A==
+-----END CERTIFICATE-----
+""")
+
+        static let tlsKey = OpenVPN.StaticKey(file: """
+# 2048 bit OpenVPN static key
+-----BEGIN OpenVPN Static key V1-----
+6acef03f62675b4b1bbd03e53b187727
+423cea742242106cb2916a8a4c829756
+3d22c7e5cef430b1103c6f66eb1fc5b3
+75a672f158e2e2e936c3faa48b035a6d
+e17beaac23b5f03b10b868d53d03521d
+8ba115059da777a60cbfd7b2c9c57472
+78a15b8f6e68a3ef7fd583ec9f398c8b
+d4735dab40cbd1e3c62a822e97489186
+c30a0b48c7c38ea32ceb056d3fa5a710
+e10ccc7a0ddb363b08c3d2777a3395e1
+0c0b6080f56309192ab5aacd4b45f55d
+a61fc77af39bd81a19218a79762c3386
+2df55785075f37d8c71dc8a42097ee43
+344739a0dd48d03025b0450cf1fb5e8c
+aeb893d9a96d1f15519bb3c4dcb40ee3
+16672ea16c012664f8a9f11255518deb
+-----END OpenVPN Static key V1-----
+""", direction: .client)!
+
+        struct Parameters {
+            let title: String
+
+            let appGroup: String
+
+            let hostname: String
+
+            let port: UInt16
+
+            let socketType: SocketType
+        }
+
+        static func make(params: Parameters) -> OpenVPN.ProviderConfiguration {
+            var builder = OpenVPN.ConfigurationBuilder()
+            builder.ca = ca
+            builder.cipher = .aes256cbc
+            builder.digest = .sha512
+            builder.compressionFraming = .compLZO
+            builder.renegotiatesAfter = nil
+            builder.remotes = [Endpoint(params.hostname, EndpointProtocol(params.socketType, params.port))]
+            builder.tlsWrap = TLSWrap(strategy: .auth, key: tlsKey)
+            builder.mtu = 1350
+            builder.routingPolicies = [.IPv4, .IPv6]
+            let cfg = builder.build()
+
+            var providerConfiguration = OpenVPN.ProviderConfiguration(params.title, appGroup: params.appGroup, configuration: cfg)
+            providerConfiguration.shouldDebug = true
+            providerConfiguration.masksPrivateData = false
+            return providerConfiguration
+        }
+    }
+}
+
+extension WireGuard {
+    struct Parameters {
+        let title: String
+
+        let appGroup: String
+
+        let clientPrivateKey: String
+
+        let clientAddress: String
+
+        let serverPublicKey: String
+
+        let serverAddress: String
+
+        let serverPort: String
+    }
+
+    struct DemoConfiguration {
+        static func make(params: Parameters) -> WireGuard.ProviderConfiguration? {
+            var builder: WireGuard.ConfigurationBuilder
+            do {
+                builder = try WireGuard.ConfigurationBuilder(params.clientPrivateKey)
+            } catch {
+                print(">>> \(error)")
+                return nil
+            }
+            builder.addresses = [params.clientAddress]
+            builder.dnsServers = ["1.1.1.1", "1.0.0.1"]
+            do {
+                try builder.addPeer(params.serverPublicKey, endpoint: "\(params.serverAddress):\(params.serverPort)")
+            } catch {
+                print(">>> \(error)")
+                return nil
+            }
+            builder.addDefaultGatewayIPv4(toPeer: 0)
+            let cfg = builder.build()
+
+            return WireGuard.ProviderConfiguration(params.title, appGroup: params.appGroup, configuration: cfg)
+        }
+    }
+}